refactor playground.py and test
